Move component specs to component directories
